@@ -1,13 +1,13 @@
 // see http://vuejs-templates.github.io/webpack for documentation.
-var path = require('path')
+var path = require('path');
 
 module.exports = {
   build: {
     env: require('./prod.env'),
-    index: path.resolve(__dirname, '../../docs/index.html'),
-    assetsRoot: path.resolve(__dirname, '../../docs'),
+    index: path.resolve(__dirname, '../../docs/examples/index.html'),
+    assetsRoot: path.resolve(__dirname, '../../docs/examples/'),
     assetsSubDirectory: 'static',
-    assetsPublicPath: '/Vue2Leaflet/',
+    assetsPublicPath: '/Vue2Leaflet/examples/',
     productionSourceMap: true,
     // Gzip off by default as many popular static hosts such as
     // Surge or Netlify already gzip all static assets for you.
@@ -29,4 +29,4 @@ module.exports = {
     // just be aware of this issue when enabling this option.
     cssSourceMap: false
   }
-}
+};
