@@ -113,8 +113,8 @@
       <l-control-layers
         :position="layersPosition"
         :collapsed="false"
-        :sortLayers="true"
-        />
+        :sort-layers="true"
+      />
       <l-tile-layer
         v-for="tileProvider in tileProviders"
         :key="tileProvider.name"
@@ -134,7 +134,7 @@
         :key="marker.id"
         :visible="marker.visible"
         :draggable="marker.draggable"
-        :lat-lng="marker.position"
+        :lat-lng.sync="marker.position"
         :icon="marker.icon"
         @click="alert(marker)">
         <l-popup :content="marker.tooltip" />
@@ -266,7 +266,7 @@ export default {
       center: [51.505, -0.09],
       opacity: 0.6,
       token: 'your token if using mapbox',
-      mapOptions: { zoomControl: false, attributionControl: false },
+      mapOptions: { zoomControl: false, attributionControl: false, zoomSnap: true },
       zoom: 13,
       minZoom: 1,
       maxZoom: 20,
