@@ -2,35 +2,48 @@
   <div>
     <div style="height: 20%; overflow: auto;">
       <h3>Simple map</h3>
-      <p>Marker is placed at {{ marker.lat }}, {{ marker.lng }}</p>
+      <p>First marker is placed at {{ withPopup.lat }}, {{ withPopup.lng }}</p>
       <p> Center is at {{ currentCenter }} and the zoom is: {{ currentZoom }} </p>
-      <button @click="showLongText">Toggle Long popup</button>
+      <button @click="showLongText">Toggle long popup</button>
     </div>
     <l-map
       :zoom="zoom"
       :center="center"
+      :options="mapOptions"
       style="height: 80%"
       @update:center="centerUpdate"
-      @update:zoom="zoomUpdate">
+      @update:zoom="zoomUpdate"
+    >
       <l-tile-layer
         :url="url"
-        :attribution="attribution"/>
-      <l-marker :lat-lng="marker">
+        :attribution="attribution"
+      />
+      <l-marker :lat-lng="withPopup">
         <l-popup>
-          <div @click="popupClick">
-            I am a tooltip
+          <div @click="innerClick">
+            I am a popup
             <p v-show="showParagraph">
               Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque sed pretium nisl, ut sagittis sapien. Sed vel sollicitudin nisi. Donec finibus semper metus id malesuada.
             </p>
           </div>
         </l-popup>
       </l-marker>
+      <l-marker :lat-lng="withTooltip">
+        <l-tooltip :options="{permanent: true, interactive: true}">
+          <div @click="innerClick">
+            I am a tooltip
+            <p v-show="showParagraph">
+              Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque sed pretium nisl, ut sagittis sapien. Sed vel sollicitudin nisi. Donec finibus semper metus id malesuada.
+            </p>
+          </div>
+        </l-tooltip>
+      </l-marker>
     </l-map>
   </div>
 </template>
 
 <script>
-import { LMap, LTileLayer, LMarker, LPopup } from 'vue2-leaflet';
+import { LMap, LTileLayer, LMarker, LPopup, LTooltip } from 'vue2-leaflet';
 
 export default {
   name: 'Example',
@@ -38,7 +51,8 @@ export default {
     LMap,
     LTileLayer,
     LMarker,
-    LPopup
+    LPopup,
+    LTooltip
   },
   data () {
     return {
@@ -46,10 +60,14 @@ export default {
       center: L.latLng(47.413220, -1.219482),
       url: 'http://{s}.tile.osm.org/{z}/{x}/{y}.png',
       attribution: '&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors',
-      marker: L.latLng(47.413220, -1.219482),
-      currentZoom: 13,
+      withPopup: L.latLng(47.413220, -1.219482),
+      withTooltip: L.latLng(47.414220, -1.250482),
+      currentZoom: 11.5,
       currentCenter: L.latLng(47.413220, -1.219482),
-      showParagraph: false
+      showParagraph: false,
+      mapOptions: {
+        zoomSnap: 0.5
+      }
     };
   },
   methods: {
@@ -62,8 +80,8 @@ export default {
     showLongText () {
       this.showParagraph = !this.showParagraph;
     },
-    popupClick () {
-      alert('Popup Click!');
+    innerClick () {
+      alert('Click!');
     }
   }
 };
