@@ -1,7 +1,6 @@
 import Vue from 'vue';
-import L from 'leaflet';
+import { L } from 'vue2-leaflet';
 import App from './App';
-
 import 'leaflet.icon.glyph';
 
 delete L.Icon.Default.prototype._getIconUrl;
