@@ -15,6 +15,14 @@ export default {
     zIndex: {
       type: Number,
       default: 1
+    },
+    tileSize: {
+      type: Number,
+      default: 256
+    },
+    noWrap: {
+      type: Boolean,
+      default: false
     }
   },
   mounted () {
@@ -22,7 +30,9 @@ export default {
       ...this.layerOptions,
       pane: this.pane,
       opacity: this.opacity,
-      zIndex: this.zIndex
+      zIndex: this.zIndex,
+      tileSize: this.tileSize,
+      noWrap: this.noWrap
     };
   }
 };
